% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\docType{class}
\name{AccumulationEvent}
\alias{AccumulationEvent}
\title{Class providing accumulation event functionality. Inherited from Event.}
\value{
Object of \code{\link{R6Class}} ManipulationEvent.
}
\description{
Provides accumulation event functionality. Inherited from Event.
}
\section{Super class}{
\code{\link[Sima:Event]{Sima::Event}} -> \code{AccumulationEvent}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-apply}{\code{AccumulationEvent$apply()}}
\item \href{#method-clone}{\code{AccumulationEvent$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="Sima" data-topic="Event" data-id="allocate">}\href{../../Sima/html/Event.html#method-allocate}{\code{Sima::Event$allocate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Sima" data-topic="Event" data-id="get_id">}\href{../../Sima/html/Event.html#method-get_id}{\code{Sima::Event$get_id()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Sima" data-topic="Event" data-id="get_name">}\href{../../Sima/html/Event.html#method-get_name}{\code{Sima::Event$get_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Sima" data-topic="Event" data-id="get_parameters">}\href{../../Sima/html/Event.html#method-get_parameters}{\code{Sima::Event$get_parameters()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Sima" data-topic="Event" data-id="initialize">}\href{../../Sima/html/Event.html#method-initialize}{\code{Sima::Event$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Sima" data-topic="Event" data-id="set_id">}\href{../../Sima/html/Event.html#method-set_id}{\code{Sima::Event$set_id()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Sima" data-topic="Event" data-id="set_parameters">}\href{../../Sima/html/Event.html#method-set_parameters}{\code{Sima::Event$set_parameters()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-apply"></a>}}
\if{latex}{\out{\hypertarget{method-apply}{}}}
\subsection{Method \code{apply()}}{
Apply an accumulation event
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccumulationEvent$apply(status, parallel = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{status}}{Current status of the simulated population.}

\item{\code{parallel}}{A logical value indicating if parallel computation is being used in simulation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link{data.table}} with the new individuals added.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccumulationEvent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
