% Generated by roxygen2 (4.0.1): do not edit by hand
\name{acp}
\alias{acp}
\title{PCA.S}
\usage{
acp(fichier, eset, bWrite = FALSE)
}
\arguments{
\item{fichier}{.}

\item{eset}{.}

\item{bWrite}{Boolean, whether to save figure in a file. Default to FALSE.}
}
\description{
Principal Component Analysis.
}

