\name{SIMA}
\alias{SIMA}
\alias{SIMA}
\docType{package}
\title{SIMA}
\description{SIMA}
\details{
\tabular{ll}{
Package: \tab SIMA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-02-18\cr
License: \tab What license is it under?\cr
}
The most important functions are: ...
}
\author{Vincent Guillemot and Justine Guegan
Maintainer: Who to complain to <yourfault@somewhere.net>}
\keyword{ package }
