\name{preambule}
\alias{preambule}
\title{Create an output arborescence in which the output files will be saved.}
\usage{
preambule(path_in = "./", path_out = "./")
}
\arguments{
  \item{path_in}{the path of the folder containing the
  input data.}

  \item{path_out}{the path of the folder where the ouput
  files should be written.}
}
\description{
Create an output arborescence in which the output files
will be saved.
}

