% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\docType{class}
\name{Simulator}
\alias{Simulator}
\title{Class providing the simulation functionality.}
\description{
Provides methods for initializing the synthetic population, parallel simulation, calibration and monitoring.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Simulator$new()}}
\item \href{#method-start_cluster}{\code{Simulator$start_cluster()}}
\item \href{#method-stop_cluster}{\code{Simulator$stop_cluster()}}
\item \href{#method-run}{\code{Simulator$run()}}
\item \href{#method-run_scenarios}{\code{Simulator$run_scenarios()}}
\item \href{#method-reconfigure}{\code{Simulator$reconfigure()}}
\item \href{#method-configure}{\code{Simulator$configure()}}
\item \href{#method-intervene}{\code{Simulator$intervene()}}
\item \href{#method-get_status}{\code{Simulator$get_status()}}
\item \href{#method-get_init}{\code{Simulator$get_init()}}
\item \href{#method-get_sample}{\code{Simulator$get_sample()}}
\item \href{#method-get_history}{\code{Simulator$get_history()}}
\item \href{#method-initialize_monitor}{\code{Simulator$initialize_monitor()}}
\item \href{#method-reset}{\code{Simulator$reset()}}
\item \href{#method-clone}{\code{Simulator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Simulator object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$new(
  initializer = NULL,
  acc_events = list(),
  man_events = list(),
  seeds = 123,
  keep_init = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initializer}}{A function that generates the initial population.}

\item{\code{acc_events}}{A list of AccumulationEvent objects.}

\item{\code{man_events}}{A list of ManipulationEvent objects.}

\item{\code{seeds}}{Seed values for random number generation.}

\item{\code{keep_init}}{A logical value indicating whether the initial status should be stored.
If \code{TRUE} the initial population can be retrieved with the method get_init.}

\item{\code{...}}{Additional arguments to \code{initializer},}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Simulator} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_cluster"></a>}}
\if{latex}{\out{\hypertarget{method-start_cluster}{}}}
\subsection{Method \code{start_cluster()}}{
Start a parallel computation cluster.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$start_cluster(
  cl,
  nc,
  export,
  packages,
  interface = c("doParallel", "doMPI")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cl}}{A cluster object.}

\item{\code{nc}}{Size of the cluster (available computing cores).}

\item{\code{export}}{A character vector giving the names of objects from the global environment that should be exported to the workers.}

\item{\code{packages}}{A character vector giving the names of packages that should be loaded on the workers.}

\item{\code{interface}}{Interface to "foreach", must match the cluster type ("doParallel" and "doMPI" are currently supported).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop_cluster"></a>}}
\if{latex}{\out{\hypertarget{method-stop_cluster}{}}}
\subsection{Method \code{stop_cluster()}}{
Stop a parallel computation cluster.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$stop_cluster()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
Progress the simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$run(N, show_progress = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{N}}{An integer indicating the length of the simulation.}

\item{\code{show_progress}}{A logical value.
If \code{TRUE} shows a progress bar that tracks the completion of the simulation
(for non-parallel computations only)}

\item{\code{unit}}{A character string giving the time unit that \code{t_sim} corresponds to.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_scenarios"></a>}}
\if{latex}{\out{\hypertarget{method-run_scenarios}{}}}
\subsection{Method \code{run_scenarios()}}{
Run various scenarios with the same initial population using different interventions with replications.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$run_scenarios(
  interventions,
  replications,
  N,
  seeds_override = NULL,
  output,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{interventions}}{A list of interventions corresponding to the scenarios to run, see 'details'.}

\item{\code{replications}}{An integer describing how many times each scenario should be replicated.}

\item{\code{N}}{A single integer or an integer vector describing the simulation runtime for each scenario.}

\item{\code{seeds_override}}{If \code{NULL}, a different seed is automatically used for each scenario.
If provided, should be an integer matrix with \code{replications} rows and \code{length(interventions)} columns.}

\item{\code{output}}{A function to apply to the population after each scenario.}

\item{\code{...}}{Additional arguments passed to \code{output}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Each element of \code{interventions} should be a named list with the following elements:
\code{acc_events}, \code{acc_pars}, \code{man_events}, \code{man_pars} and \code{pars}.
The element \code{acc_events} should be a list of objects of class \code{AccumulationEvent} and
\code{acc_pars} should be a list of the same length of numeric vectors giving the parameter values for these events.
Similarly, \code{man_events} should be a list of object of class \code{ManipulationEvent} with \code{man_pars}
being a list of the same length of numeric vectors giving the parameter values.
Finally, \code{pars} should be a list of parameter values (see \code{configure}).
For additional details, see \code{intervene}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reconfigure"></a>}}
\if{latex}{\out{\hypertarget{method-reconfigure}{}}}
\subsection{Method \code{reconfigure()}}{
Configure the parameters of Event objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$reconfigure(pars)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pars}}{A named list of parameters, see 'details'.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The parameter list \code{pars} should be structured as follows:
Each element should have a name of an Event object and each
element should be a named list where the names correspond
to the names of the parameters of whose values are to be configured.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-configure"></a>}}
\if{latex}{\out{\hypertarget{method-configure}{}}}
\subsection{Method \code{configure()}}{
Configure events, run a simulation and compute a summary statistic for calibration.
The original status is restored after completion.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$configure(N, pars, output, seeds = 123, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{N}}{A numeric value indicating the length of the simulation}

\item{\code{pars}}{A list of parameter values, see 'details'.}

\item{\code{output}}{A function that computes a summary statistic from the simulation.}

\item{\code{seeds}}{Seed values for random number generation.}

\item{\code{...}}{Additional arguments passed to \code{output}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The parameter list \code{pars} should be structured as follows:
Each element should have a name of an Event object and each
element should be a named list where the names correspond
to the names of the parameters of whose values are to be configured.
}

\subsection{Returns}{
The output of \code{output_function} evaluated for the resulting population.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-intervene"></a>}}
\if{latex}{\out{\hypertarget{method-intervene}{}}}
\subsection{Method \code{intervene()}}{
Carry out an intervention that applies a set of accumulation events and manipulation events,
and changes the values of specific parameters.
By default, the events are applied first, then the parameters are changed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$intervene(
  acc_events = list(),
  man_events = list(),
  pars = list(),
  events_first = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{acc_events}}{A list of accumulation events to apply.}

\item{\code{man_events}}{A list of manipulation events to apply.}

\item{\code{pars}}{A list of parameter values to set, see 'details'.}

\item{\code{events_first}}{A logical value. If \code{TRUE} (the default) then the events are applied before any changes to the parameters. If \code{FALSE} the parameter values are changed first.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The parameter list \code{pars} should be structured as follows:
Each element should have a name of an Event object and each
element should be a named list where the names correspond
to the names of the parameters of whose values are to be configured.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_status"></a>}}
\if{latex}{\out{\hypertarget{method-get_status}{}}}
\subsection{Method \code{get_status()}}{
Get the status of the current population.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$get_status()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.table of the current status.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_init"></a>}}
\if{latex}{\out{\hypertarget{method-get_init}{}}}
\subsection{Method \code{get_init()}}{
Get the initial status of the population, if available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$get_init()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.table of the initial status if it was recorded, otherwise NULL.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_sample"></a>}}
\if{latex}{\out{\hypertarget{method-get_sample}{}}}
\subsection{Method \code{get_sample()}}{
.
Get a sample of the current population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$get_sample(sampler, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sampler}}{A function that constructs the output sample.}

\item{\code{...}}{Additional arguments passed to \code{sampler}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The sample generated by \code{sampler}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_history"></a>}}
\if{latex}{\out{\hypertarget{method-get_history}{}}}
\subsection{Method \code{get_history()}}{
Get the simulation history as recorded by the monitor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$get_history()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
If the monitor is enabled, A list with an element for each evaluation of the monitor aggregators at their specified intervals
These elements are lists as well, where the first element is the time index, and the remaining elements
correspond to the values returned by the evaluated aggregators of the monitor. If no monitor is specified, returns NULL (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initialize_monitor"></a>}}
\if{latex}{\out{\hypertarget{method-initialize_monitor}{}}}
\subsection{Method \code{initialize_monitor()}}{
Initialize monitors to track specific status variables throughout the simulation at certain intervals.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$initialize_monitor(aggregators, intervals, combiners = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aggregators}}{A list of functions to compute aggregate summary statistics from the simulated population.}

\item{\code{intervals}}{A numeric vector giving the time interval after which each aggregator should be evaluated.}

\item{\code{combiners}}{If parallel computation is used, this is a list of functions that combine the aggregator results from each parallel worker.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
Reset the simulated population to the initial state (if initial status was kept).
Note that any changes to event parameters are kept as is and have to be configured manually, if so desired.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$reset(force = FALSE, seeds = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{force}}{If \code{TRUE}, reset the simulator even if the initial status is not available.}

\item{\code{seeds}}{An integer vector. If given, sets the pseudo-RNG seeds.
This will then reset the seed numbers, and the simulation time.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
