% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\docType{class}
\name{Simulator}
\alias{Simulator}
\title{Class providing the simulation functionality.}
\description{
Provides methods for initializing the synthetic population, parallel simulation, calibration and monitoring.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Simulator$new()}}
\item \href{#method-start_cluster}{\code{Simulator$start_cluster()}}
\item \href{#method-stop_cluster}{\code{Simulator$stop_cluster()}}
\item \href{#method-run}{\code{Simulator$run()}}
\item \href{#method-reconfigure}{\code{Simulator$reconfigure()}}
\item \href{#method-configure}{\code{Simulator$configure()}}
\item \href{#method-get_population}{\code{Simulator$get_population()}}
\item \href{#method-get_sample}{\code{Simulator$get_sample()}}
\item \href{#method-initialize_monitor}{\code{Simulator$initialize_monitor()}}
\item \href{#method-clone}{\code{Simulator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Simulator object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$new(
  initializer = NULL,
  manipulation_events = list(),
  accumulation_events = list(),
  time = Time$new("day", Sys.Date()),
  n,
  seeds = 123
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initializer}}{A function to generate the initial population.}

\item{\code{manipulation_events}}{A list of ManipulationEvent objects.}

\item{\code{accumulation_events}}{A list of AccumulationEvent objects.}

\item{\code{time}}{A Time object.}

\item{\code{n}}{Size of the initial population.}

\item{\code{seeds}}{Seed values for random number generation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{\link{Simulator}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_cluster"></a>}}
\if{latex}{\out{\hypertarget{method-start_cluster}{}}}
\subsection{Method \code{start_cluster()}}{
Start a parallel computation cluster.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$start_cluster(cl, nc, export, packages, interface = "doParallel")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cl}}{A cluster object.}

\item{\code{nc}}{Size of the cluster (available computing cores).}

\item{\code{export}}{A character vector giving the names of objects from the global environment that should be exported to the workers.}

\item{\code{packages}}{A character vector giving the names of packages that should be loaded on the workers.}

\item{\code{interface}}{Interface to "foreach", must match the cluster type ("doParallel" and "doMPI" are currently supported).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop_cluster"></a>}}
\if{latex}{\out{\hypertarget{method-stop_cluster}{}}}
\subsection{Method \code{stop_cluster()}}{
Stop a parallel computation cluster.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$stop_cluster()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
Progress the simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$run(t_sim)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t_sim}}{A numeric value indicating the length of time to process.}

\item{\code{unit}}{A character string giving the time unit that \code{t_sim} corresponds to.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reconfigure"></a>}}
\if{latex}{\out{\hypertarget{method-reconfigure}{}}}
\subsection{Method \code{reconfigure()}}{
Configure the parameters of Event objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$reconfigure(p)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p}}{A named list of parameters, see 'details'.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The parameter list \code{p} should be structured as follows:
Each element should be named with a name of an event object.
Each element should be a list with two elements, one with
the parameters values and the second of equal length
giving the indices which to replace.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-configure"></a>}}
\if{latex}{\out{\hypertarget{method-configure}{}}}
\subsection{Method \code{configure()}}{
Configure events, run a simulation and compute a summary statistic for calibration. The original status is restored after completion.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$configure(t_sim, p, output, seeds = 123, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t_sim}}{A numeric value indicating the length of time to process.}

\item{\code{p}}{A list of parameter values, see 'details'.}

\item{\code{output}}{A function that computes a summary statistic from the simulation.}

\item{\code{seeds}}{Seed values for random number generation.}

\item{\code{...}}{Additional arguments passed to \code{output}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The parameter list \code{p} should be structured as follows:
Each element should be named with a name of an event object.
Each element should be a list with two elements, one with
the parameters values and the second of equal length
giving the indices which to replace.
}

\subsection{Returns}{
The output of \code{output_function} evaluated for the resulting population.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_population"></a>}}
\if{latex}{\out{\hypertarget{method-get_population}{}}}
\subsection{Method \code{get_population()}}{
Get the status of the current population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$get_population()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.table of the current status
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_sample"></a>}}
\if{latex}{\out{\hypertarget{method-get_sample}{}}}
\subsection{Method \code{get_sample()}}{
Get a sample of the current population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$get_sample(sampler, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sampler}}{A function that constructs the output sample.}

\item{\code{...}}{Additional arguments passed to \code{sampler}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The sample generated by \code{sampler}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initialize_monitor"></a>}}
\if{latex}{\out{\hypertarget{method-initialize_monitor}{}}}
\subsection{Method \code{initialize_monitor()}}{
Initialize monitors to track specific status variables throughout the simulation at certain intervals.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$initialize_monitor(aggregators, intervals, combiners = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aggregators}}{A list of functions to compute aggregate summary statistics from the simulated population.}

\item{\code{intervals}}{A numeric vector giving the time interval after which each aggregator should be evaluated.}

\item{\code{combiners}}{If parallel computation is used, this is a list of functions that combine the aggregator results from each parallel worker.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
